+++ src/relayFeeCalculator/chain-queries/zksync.ts
@@ -1,8 +1,8 @@
 import { getDeployedAddress } from "../../utils/DeploymentUtils";
 import { DEFAULT_LOGGER, Logger } from "../relayFeeCalculator";
 import { providers } from "ethers";
-import { TOKEN_SYMBOLS_MAP, CHAIN_IDs } from "../../constants";
-import QueryBase, { DEFAULT_SIMULATED_RELAYER_ADDRESS } from "./baseQuery";
+import { TOKEN_SYMBOLS_MAP, CHAIN_IDs, DEFAULT_SIMULATED_RELAYER_ADDRESS } from "../../constants";
+import QueryBase from "./baseQuery";
 
 export class ZkSyncQueries extends QueryBase {
   constructor(
